requirements.txt
django
django-cart
django-stdimage == 1.2.2
django-tables2
django-filter==0.9.2
django-filters==0.1.0
transliterate